+++ zap_dev/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2-dev
+
+- Members defined in a `<script context="module">` tag are now visible to
+  components importing the file containing the script.
+
 ## 0.1.1
 
 - Emit static HTML for fragments known not to change, decreasing code size.